Change license to the Mozilla Public License, version 2.0
Why am I relicensing to MPL v2? I have a range of improvements that I'll be making to
the library for use in a project I'm working on. I think it's fair that people be able to use
these improvements in their applications, even for commercial applications, and that any
improvements they make are shared back so everyone else can make use of them.

The MPL v2 allows for static linking, something that is typically done for embedded applications,
while also ensuring that the source code must be disclosed when the binary
is distributed.