Help ensure that operation matches build configuration by adding flags and a return value to httpdInit() and httpdInitEx()
Previously one could initialize httpd with the https port but forget to enable ssl support, potentially risking
hosting an http server when an https server was expected, a potential security issue. Altering the API to include the intent
via flags reduces this security risk. Introduce a return value to report a mismatched condition to the user.

Modify FreeRTOS httpdPlatInit() to check that if SSL is enabled that the SSL feature is enabled.