httpd-freertos.c - Fix data stuck in SSL connection issue
Observed behavior was that during some data transfers over SSL the transfer would not complete.

Upon measurement it became apparent that there were some number of bytes stuck inside of SSL.

It looks like select(), both under lwip and Linux, can result in a case where the select() returns, SSL_read() returns bytes
but not all of the bytes and then the subsequent call to select() does not result in the select() returning immediately due to
these pending bytes.

Furthermore, SSL_pending(), called after the SSL_read(), indicates that these bytes are available to read.

This change causes SSL_read()/SSL_pending() to be called in a loop until there is no more data indicated as pending in SSL
at which time the loop exits and select() is called again. This isn't a great fix but I can't figure out why select()
isn't returning for the pending data or why the implemention isn't working as expected. If you know please drop me
an email.