Re-comment noisy OTA upload log message (#9)
This line is commented out in the latest revision of the upstream project. When it is left in, it generates an extremely large amount of log messages during OTA upload, slowing the whole upload process down. It should be commented (or otherwise disabled) unless one is debugging a low level of the upload process.