Fix issue #38, invalid multipartBoundary (#40)
* Renamed base64_decode and base64_encode because they can collied with symbols in ESP-IDF if WPA functions like esp_wifi_sta_wpa2_ent_set_identity are used (see esp32.rom.ld). This can lead to the wrong address being called and a wrong result.

* Changed to pragma once style

* Breaking change in multipartBoundary, fixed header manipulation

Previously the multipartBoundary pointer was used to change the original header to include the two required hyphens but that manipulated the original header buffer. This manipulation was removed but the pointer moved by a wrong amount. With this commit multipartBoundary will always point to the start of the boundary value but users need to prepend the two hyphens manually.